+++ app/frontend/src/features/digital-object-imports/components/digital-object-imports-list.tsx
@@ -1,10 +1,11 @@
 import { Link, useSearchParams } from 'react-router';
-import type { ColumnDef, PaginationState, Updater } from '@tanstack/react-table';
+import type { ColumnDef } from '@tanstack/react-table';
 import TableBuilder from '@/components/ui/table-builder/table-builder';
 import { columnDefs } from '../utils/digital-object-imports-column-defs';
 import { useDigitalObjectImportsSuspenseQuery } from '../api/get-digital-object-imports';
 import { DigitalObjectImportSummary } from '@/types/api';
 import { StatusFilter } from './status-filter';
+import { useTablePagination } from '@/hooks/use-table-pagination';
 
 interface DigitalObjectImportsListProps {
   importJobId: string;
@@ -14,26 +15,11 @@ const options = ['pending', 'success', 'failure'];
 
 export const DigitalObjectImportsList = ({ importJobId }: DigitalObjectImportsListProps) => {
   const [searchParams, setSearchParams] = useSearchParams();
-  const page = Number(searchParams.get('page')) || 1;
   const status = searchParams.get('status') ?? undefined;
+  const { page, getPaginationProps } = useTablePagination();
 
   const query = useDigitalObjectImportsSuspenseQuery({ importJobId, page, status });
   const { digitalObjectImports, pagination, importJobName } = query.data;
-
-  // TODO: This is the same logic as in import-jobs-list.tsx; might work well as a custom hook
-  const paginationState: PaginationState = {
-    pageIndex: page - 1,
-    pageSize: pagination.perPage,
-  };
-
-  // TODO: This is the same logic as in import-jobs-list.tsx; might work well as a custom hook
-  const handlePaginationChange = (updater: Updater<PaginationState>) => {
-    const next = typeof updater === 'function' ? updater(paginationState) : updater;
-    setSearchParams((prev) => {
-      prev.set('page', String(next.pageIndex + 1));
-      return prev;
-    });
-  };
 
   const handleStatusChange = (newStatus: string | null) => {
     setSearchParams((prev) => {
@@ -69,11 +55,7 @@ export const DigitalObjectImportsList = ({ importJobId }: DigitalObjectImportsLi
         data={digitalObjectImports}
         columns={columnDefs as ColumnDef<DigitalObjectImportSummary>[]}
         size="sm"
-        pagination={{
-          state: paginationState,
-          onPaginationChange: handlePaginationChange,
-          rowCount: pagination.totalCount,
-        }}
+        pagination={getPaginationProps(pagination)}
       />
     </div>
   );
