+++ app/frontend/src/features/export-jobs/api/download-csv-export.tsx
@@ -1,0 +1,4 @@
+import { BASE_URL } from '@/lib/api-client';
+
+export const getCsvExportDownloadUrl = (exportJobId: number) =>
+  `${BASE_URL}/export_jobs/${exportJobId}/download`;
